==> 3_Check_M's result.ipynb <==
%%time
from concurrent.futures import ProcessPoolExecutor
from datetime import datetime, timedelta
import matplotlib.pyplot as plt
import numpy as np
import os
import pandas as pd
import random
from statistics import mean 
import time
import torch

import sys
script_path = os.path.abspath("pyscripts")
if script_path not in sys.path:
    sys.path.append(script_path)
from utils import *
from load_data import *
from lstm import *

# time sequence
train_window = 48
input_size = 1
output_size = 24
hidden_size = 48
num_layers = 2
learning_rate = 0.005

# use GPU if available
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

obs = np.load('/home/graduate/fbx5002/disk10TB/IEE/IoT_Weather_Data/Geotab_Obs_Ratio05.npy')
# swap rows and columns, obs = np.swapaxes(obs, 0, 1)
obs = obs.T
times = np.load('/home/graduate/fbx5002/disk10TB/IEE/IoT_Weather_Data/Geotab_Obs_TS_Ratio05.npy')
times = [datetime.utcfromtimestamp(x.tolist()/1e9) for x in times]
stations = pd.read_csv('/home/graduate/fbx5002/disk10TB/IEE/IoT_Weather_Data/selected_Geotab_stations_Ratio05.csv')
stations = stations[stations.columns[[3,1,2,4]]]

weather_station_2015_2020 = pd.read_csv("/home/graduate/fbx5002/disk10TB/IEE/IoT_Weather_Data/weather_station_2015_2020.csv")
weather_time_2015_2020 = np.load('/home/graduate/fbx5002/disk10TB/IEE/IoT_Weather_Data/weather_time_2015_2020.npy', allow_pickle=True)

weather_obs_2015_2020 = np.load('/home/graduate/fbx5002/disk10TB/IEE/IoT_Weather_Data/weather_obs_2015_2020.npy')
weather_obs_2010_2020 = np.load('/home/graduate/fbx5002/disk10TB/IEE/IoT_Weather_Data/weather_obs_2010_2020.npy')

stations5 = set(stations.sort_values('missing')[:455]['geohash'])
stations10= set(stations.sort_values('missing')[:1650]['geohash'])
stations15= set(stations.sort_values('missing')[:2107]['geohash'])
stations20= set(stations.sort_values('missing')[:2562]['geohash'])
wu_stations = set(weather_station_2015_2020['station'])

num_epochs = 101
epoch_interval = 2

# two dataset used different norm_min, norm_max
# read save data
loaded5 = torch.load('./checkpoint/IoT_final/TensorData/Torch_db5WU')
train_tensors5 = loaded5['train']
test_tensors5 = loaded5['test']
test_times5 = loaded5['selected_days']
norm_min5, norm_max5 = train_tensors5.min, train_tensors5.max
print(norm_min5, norm_max5)

# read save data
loaded10 = torch.load('./checkpoint/IoT_final/TensorData/Torch_db10WU')
train_tensors10 = loaded10['train']
test_tensors10 = loaded10['test']
test_times10 = loaded10['selected_days']
norm_min10, norm_max10 = train_tensors10.min, train_tensors10.max
print(norm_min10, norm_max10)

checkpoint5 = torch.load('./checkpoint/IoT_final/checkpoint-5.5-{epoch}.pth.tar'.format(epoch=60))
checkpoint10 = torch.load('./checkpoint/IoT_final/checkpoint-10-{epoch}.pth.tar'.format(epoch=23))
checkpoint15 = torch.load('./checkpoint/IoT_final/checkpoint-15-{epoch}.pth.tar'.format(epoch=8))
checkpoint20 = torch.load('./checkpoint/IoT_final/checkpoint-20-{epoch}.pth.tar'.format(epoch=7))

checkpoint5IoT5WU = torch.load('./checkpoint/IoT_final/checkpoint-5.5-IoT-5-W-{epoch}.pth.tar'.format(epoch=18))
checkpoint10IoT5WU = torch.load('./checkpoint/IoT_final/checkpoint-10-IoT-5-W-{epoch}.pth.tar'.format(epoch=9))
checkpoint15IoT5WU = torch.load('./checkpoint/IoT_final/checkpoint-15-IoT-5-W-{epoch}.pth.tar'.format(epoch=5))
checkpoint20IoT5WU = torch.load('./checkpoint/IoT_final/checkpoint-20-IoT-5-W-{epoch}.pth.tar'.format(epoch=7))

checkpoint5IoT10WU = torch.load('./checkpoint/IoT_final/checkpoint-5.5-IoT-10-W-{epoch}.pth.tar'.format(epoch=15))
checkpoint10IoT10WU = torch.load('./checkpoint/IoT_final/checkpoint-10-IoT-10-W-{epoch}.pth.tar'.format(epoch=12))
checkpoint15IoT10WU = torch.load('./checkpoint/IoT_final/checkpoint-15-IoT-10-W-{epoch}.pth.tar'.format(epoch=5))
checkpoint20IoT10WU = torch.load('./checkpoint/IoT_final/checkpoint-20-IoT-10-W-{epoch}.pth.tar'.format(epoch=3))


_, model5, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)
_, model10, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)
_, model15, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)
_, model20, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)

_, model5IoT5WU, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)
_, model10IoT5WU, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)
_, model15IoT5WU, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)
_, model20IoT5WU, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)

_, model5IoT10WU, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)
_, model10IoT10WU, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)
_, model15IoT10WU, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)
_, model20IoT10WU, _ = initial_model(input_size, output_size, hidden_size, num_layers, learning_rate, device)


model5.load_state_dict(checkpoint5['bestLossModel'])
model10.load_state_dict(checkpoint10['bestLossModel'])
model15.load_state_dict(checkpoint15['bestLossModel'])
model20.load_state_dict(checkpoint20['bestLossModel'])

model5IoT5WU.load_state_dict(checkpoint5IoT5WU['bestLossModel'])
model10IoT5WU.load_state_dict(checkpoint10IoT5WU['bestLossModel'])
model15IoT5WU.load_state_dict(checkpoint15IoT5WU['bestLossModel'])
model20IoT5WU.load_state_dict(checkpoint20IoT5WU['bestLossModel'])

model5IoT10WU.load_state_dict(checkpoint5IoT10WU['bestLossModel'])
model10IoT10WU.load_state_dict(checkpoint10IoT10WU['bestLossModel'])
model15IoT10WU.load_state_dict(checkpoint15IoT10WU['bestLossModel'])
model20IoT10WU.load_state_dict(checkpoint20IoT10WU['bestLossModel'])

test_pred_orig_dict5 = predict(model5, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)
test_pred_orig_dict10 = predict(model10, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)
test_pred_orig_dict15 = predict(model15, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)
test_pred_orig_dict20 = predict(model20, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)

test_pred_orig_dict5_5WU = predict(model5IoT5WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)
test_pred_orig_dict10_5WU = predict(model10IoT5WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)
test_pred_orig_dict15_5WU = predict(model15IoT5WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)
test_pred_orig_dict20_5WU = predict(model20IoT5WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)

test_pred_orig_dict5_10WU = predict(model5IoT10WU, test_tensors10, (norm_min10, norm_max10), device, stationlist = stations5)
test_pred_orig_dict10_10WU = predict(model10IoT10WU, test_tensors10, (norm_min10, norm_max10), device, stationlist = stations5)
test_pred_orig_dict15_10WU = predict(model15IoT10WU, test_tensors10, (norm_min10, norm_max10), device, stationlist = stations5)
test_pred_orig_dict20_10WU = predict(model20IoT10WU, test_tensors10, (norm_min10, norm_max10), device, stationlist = stations5)

# Here, let's try to use a model to predict same observations normalized in other values different from train_tensors
# test_pred_orig_dict5_10WU = predict(model5IoT10WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)
# test_pred_orig_dict10_10WU = predict(model10IoT10WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)
# test_pred_orig_dict15_10WU = predict(model15IoT10WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)
# test_pred_orig_dict20_10WU = predict(model20IoT10WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations5)

##################################
_,_,_,_ = stat_scores(test_pred_orig_dict5)
_,_,_,_ = stat_scores(test_pred_orig_dict10)
_,_,_,_ = stat_scores(test_pred_orig_dict15)
_,_,_,_ = stat_scores(test_pred_orig_dict20)
print('\n')
_,_,_,_ = stat_scores(test_pred_orig_dict5_5WU)
_,_,_,_ = stat_scores(test_pred_orig_dict10_5WU)
_,_,_,_ = stat_scores(test_pred_orig_dict15_5WU)
_,_,_,_ = stat_scores(test_pred_orig_dict20_5WU)
print('\n')
_,_,_,_ = stat_scores(test_pred_orig_dict5_10WU)
_,_,_,_ = stat_scores(test_pred_orig_dict10_10WU)
_,_,_,_ = stat_scores(test_pred_orig_dict15_10WU)
_,_,_,_ = stat_scores(test_pred_orig_dict20_10WU)

test_pred_orig_dict10 = predict(model10, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations10)
test_pred_orig_dict10_5WU = predict(model10IoT5WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations10)
test_pred_orig_dict10_10WU = predict(model10IoT10WU, test_tensors10, (norm_min10, norm_max10), device, stationlist = stations10)

test_pred_orig_dict15 = predict(model15, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations15)
test_pred_orig_dict15_5WU = predict(model15IoT5WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations15)
test_pred_orig_dict15_10WU = predict(model15IoT10WU, test_tensors10, (norm_min10, norm_max10), device, stationlist = stations15)

test_pred_orig_dict20 = predict(model20, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations20)
test_pred_orig_dict20_5WU = predict(model20IoT5WU, test_tensors5, (norm_min5, norm_max5), device, stationlist = stations20)
test_pred_orig_dict20_10WU = predict(model20IoT10WU, test_tensors10, (norm_min10, norm_max10), device, stationlist = stations20)

##############################
_,_,_,_ = stat_scores(test_pred_orig_dict10)
_,_,_,_ = stat_scores(test_pred_orig_dict10_5WU)
_,_,_,_ = stat_scores(test_pred_orig_dict10_10WU)
print('\n')
_,_,_,_ = stat_scores(test_pred_orig_dict15)
_,_,_,_= stat_scores(test_pred_orig_dict15_5WU)
_,_,_,_ = stat_scores(test_pred_orig_dict10_10WU)
print('\n')
_,_,_,_= stat_scores(test_pred_orig_dict20)
_,_,_,_= stat_scores(test_pred_orig_dict20_5WU)
_,_,_,_ = stat_scores(test_pred_orig_dict10_10WU)
